refactor(rest): extract procedural-models routes into routes/procedural_models.py
Moves the viewer cellbuilder's /api/scopes/{scope}/procedural-models* and
procedural-templates document CRUD, compile/preview/export/import job
enqueues, compile-log/stats readers, and the equipment/system catalog
sync/resync writers out of the create_app closure into a standalone
APIRouter, following the pattern set by routes/plugins.py.

Lifts require_catalog_pool, audit_event (as RestContext.audit),
advertised_engine_capability onto routes/deps.py so the new module reaches
storage/queue/db through RestContext instead of closure variables; keeps
_audit/_require_catalog_pool aliases bound in create_app for the routes
still inside it. Registered right after plugins_router, preserving the
route order fixed-segment catalog listings and equipment-types/*sync must
win before procedural-models/{model_id}.
